Moved DeleteParcel tests
Also removed ParcelsApiTest.cs file
